[chore] made nav component sticky at bottom of screen so content can be scrolled behind it
