+++ src/app/(with-nav)/view-logs/page.tsx
@@ -1,9 +1,22 @@
-import AllStudyLogs from '@/Components/AllStudyLogs';
+import StudyLogEntry from '@/Components/StudyLogEntry';
+import { getAllStudyLogs } from '@/lib/data';
 
-export default function ViewAllLogsPage() {
+const getData = async () => {
+  const data = await getAllStudyLogs();
+  return data;
+};
+
+export default async function ViewAllLogsPage() {
+  const studyLogs = await getData();
+
   return (
     <div className='flex flex-col gap-4'>
-      <AllStudyLogs />
+      {studyLogs.map((log) => (
+        <StudyLogEntry key={log.id} {...log} />
+      ))}
+      {studyLogs.map((log) => (
+        <StudyLogEntry key={log.id} {...log} />
+      ))}
     </div>
   );
 }
